flask_deep/darkflow_yolo/requirements.txt
numpy
matplotlib
cython
opencv-python
tensorflow-gpu
icrawler
lxml
